requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==4.1.1
celery==5.2.7
certifi==2020.4.5.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-celery-beat==2.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-action-serializer==1.0.5
drf-writable-nested==0.6.4
drf-yasg==1.20.0
dropbox==11.33.0
facebook-sdk==3.1.0
Faker==4.1.0
fcm-django==1.0.12
firebase-admin==5.2.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.7.3
google-cloud-firestore==2.1.3
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.52.0
grpcio==1.47.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.2.4
MarkupSafe==1.1.1
msgpack==1.0.4
oauthlib==3.1.0
packaging==20.4
Pillow==9.2.0
ply==3.11
prompt-toolkit==3.0.30
proto-plus==1.22.0
protobuf==3.19.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.1
python-decouple==3.6
python-twitter==3.5
python3-openid==3.2.0
pytz==2021.3
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
stone==3.3.1
text-unidecode==1.3
toml==0.10.1
twilio==7.12.1
tzdata==2022.2
uritemplate==3.0.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
xlwt==1.3.0
